Create useGoogleSignIn custom hook
